requirements.txt
python-box==7.1.1
wandb==0.16.0
torchtext==0.6.0
pyyaml==6.0.1
Pillow==10.1.0
numpy==1.19.3
pandas==1.4.4
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
scikit-learn==1.3.2
scipy==1.9.2
datasets==2.18.0
transformers==4.38.2

pip install torch==2.0.0 torchvision==0.15.1 torchaudio==2.0.1 --index-url https://download.pytorch.org/whl/cu118